live_monitor-main/services/analyzer/requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0

# Performance
tqdm>=4.66.0

# Optional: for future features
redis>=4.6.0
pyarrow>=12.0.0  # for parquet support

matplotlib
minio
